Avoid power-saving-sleep races against our child process closing.
If we invoke window.close() after having released our wake-locks, it is
possible that the device will go into a deep sleep before the content
process cleans up.  This may result in a race where when mozAlarms next
wakes the device up, the cleanup process will race the queueing of the
system message to wake the email app up.  SystemMessagesInternal may see
the unregistering event of the email app and decide to not try and spin up
the email service.

This scenario ends up permanently breaking sync under the requestsync API
because it checks if there are already pending messages, but things will
probably work the next time under mozAlarms since it will always send the
message.  (Although I haven't fully traced it.)